+++ CHANGELOG.md
@@ -6,6 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v2.0.0] - 2021-12-23
+### Added
+- PHP8.1 support
+- more configuration methods
+- RateLimit information to RateLimitException
+
+### Removed
+PHP7.4 support
+
 ## [v1.0.0] - 2021-05-12
 ### Added
 - Initial release